requirements.txt
altgraph==0.17
asgiref==3.2.5
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cycler==0.10.0
Django==3.0.4
django-filter==2.4.0
djangorestframework==3.12.2
express==1.0
future==0.18.2
idna==2.10
isort==5.6.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
pefile==2019.4.18
plot==0.6.5
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.7.4
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.25.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
typing==3.7.4.1
urllib3==1.26.2
webdriver-manager==3.2.2
wrapt==1.12.1
